Solves STR #2428: Silence float->int conversion warnings
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@7835 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Quartz_Printer.mm
@@ -164,8 +164,8 @@ int Fl_Printer::printable_rect(int *w, int *h)
   
   x = (int)pmRect.left;
   y = (int)pmRect.top;
-  *w = (int)(pmRect.right - x) / scale_x + 1;
-  *h = (int)(pmRect.bottom - y) / scale_y + 1;
+  *w = int((int)(pmRect.right - x) / scale_x + 1);
+  *h = int((int)(pmRect.bottom - y) / scale_y + 1);
   return 0;
 }
